add auth to delete route

+++ Project 3/TodoApp/routers/todos.py
@@ -47,8 +47,11 @@ async def read_all(user: user_dependency, db: db_dependency):
     return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 @router.get("/todo/{todo_id}")
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
     raise(HTTPException(status_code=404, detail="Todo not found"))
@@ -57,6 +60,7 @@ async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
 async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
     if user is None:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+
     #note we add the owner_id as the model for Todos includes it, without this we raise the exception and cannot post the todo
     todo_model=Todos(**todo_request.model_dump(), owner_id=user.get('id'))
     # get db ready here
@@ -68,6 +72,7 @@ async def create_todo(user: user_dependency, db: db_dependency, todo_request: To
 async def update_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest, todo_id: int =Path(gt=0)):
     if user is None:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+
     todo_model = db.query(Todos).filter(Todos.id == todo_id).\
         filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
@@ -81,11 +86,14 @@ async def update_todo(user: user_dependency, db: db_dependency, todo_request: To
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency ,db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
 
     db.commit()
